Created configIMU.h & .cpp
configIMU.h: Added IMU ADDRESS, some resgiters to start it's
configuration and a first function to initialize the acelerometer and
gyroscope.

configIMU.cpp: Added a first version of a function that initialize the
acelerometer and gyroscope. Also, it takes the IMU ADDRESS and sent it
to the IMU via I2C.